implement parallel processing
This could possibly be improved. For now it uses a custom infinite
loop and does periodical polling. The overhead should be marginal.

This solution has the advantage of easy error handling and cancelling
of other subprocesses, if one fails.

Other solutions that use threading (multiprocessing, concurrency.futures)
are complicated to handle in that regard.