assemble: implement multithreading
Note: This is behaving essentially different to the createMotifs thread.
While there the nstruct is the total number and is distributed among the
threads, here each thread is passed the total nstruct number.

Reasoning: rosetta uses a score filter and discards structures that fail.
So starting many threads and merging the results might yield different
results.

TODO: Maybe this doesn't matter at all and we can implement it the same
way as with createMotifs?