extract: cd to temp dir when running rna_extract
Rosetta rna_extract does not have an option to specify an output
directory and simply dumps everything into the working directory.
This makes it impossible to run multiple jobs in parallel, because
the S_00000x.pdb files would interfere with each other.

To fix it run rna_extract commands in the associated constraints
temp directory.