simulation: use 'rmsd_native', fixup for c038e7f52196f73992075b9606605a3b6c42f3f4
Contrary to the commit message, the new native rmsd values were
store as 'rmsd_native'. Leave this for now because the old ones
were stored as nm instead of angstrom, so they weren't compatible
anways.

In the future simply re-evaluate all the old predictions using the
new rosetta clustering and rmsd valuation and get rid of the
parallel workflow.