simulation: drop compare method, refactor print_status, create get_status
Native comparison is now done using a compare option to
print_status(). This makes it possible to show everything in one
nice table.

This also introduces a new get_status() method to retrieve all sorts
of information about a cst as a dict. This will be useful as an
API for the web interface, but needs to be extended first.

Another method get_csts() is added to get a list of all constraints
sets of the prediction.